pyproject.toml
[tool.poetry]
name = "pre-commit-hooks"
version = "0.1.0"
description = "Handy pre-commit hooks used at Prophecy"
authors = ["Scott Lyons <scott@prophecy.io>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Markdown = "^3.3.7"
markdown-full-yaml-metadata = "^2.1.0"
GitPython = "^3.1.27"
typer = "^0.6.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pre-commit = "^2.19.0"

[tool.poetry.scripts]
check-docusaurus-frontmatter = "pre_commit_hooks.check_docusaurus_frontmatter:main"
commit-message-reuse = "pre_commit_hooks.commit_message_reuse:main"
word-veto = "pre_commit_hooks.word_veto:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
